Refactor jinja organization for ease of reuse
