+++ trestle/core/jinja/tags.py
@@ -13,18 +13,18 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Trestle utilities to customize ."""
+"""Trestle utilities to customize jinja tags."""
 import logging
 from datetime import date
 
 import frontmatter
 
-from jinja2 import lexer, nodes
+from jinja2 import lexer
 from jinja2.environment import Environment
-from jinja2.ext import Extension
 from jinja2.parser import Parser
 
 from trestle.common import err
+from trestle.core.jinja.base import TrestleJinjaExtension
 from trestle.core.markdown import docs_markdown_node, markdown_const
 
 logger = logging.getLogger(__name__)
@@ -41,36 +41,6 @@ def adjust_heading_level(input_md: str, expected: int) -> str:
             mdn.change_header_level_by(delta)
             output_md = mdn.content.raw_text
     return output_md
-
-
-class TrestleJinjaExtension(Extension):
-    """Class to define common methods to be inherited from for use in trestle."""
-
-    # This
-    max_tag_parse = 20
-
-    def __init__(self, environment: Environment) -> None:
-        """Ensure enviroment is set and carried into class vars."""
-        super().__init__(environment)
-
-    @staticmethod
-    def parse_expression(parser):
-        """Safely parse jinja expression."""
-        # Licensed under MIT from:
-        # https://github.com/MoritzS/jinja2-django-tags/blob/master/jdj_tags/extensions.py#L424
-        # Due to how the jinja2 parser works, it treats "foo" "bar" as a single
-        # string literal as it is the case in python.
-        # But the url tag in django supports multiple string arguments, e.g.
-        # "{% url 'my_view' 'arg1' 'arg2' %}".
-        # That's why we have to check if it's a string literal first.
-        token = parser.stream.current
-        if token.test(lexer.TOKEN_STRING):
-            expr = nodes.Const(token.value, lineno=token.lineno)
-            next(parser.stream)
-        else:
-            expr = parser.parse_expression(False)
-
-        return expr
 
 
 class MDSectionInclude(TrestleJinjaExtension):
